Handle memoize edge-case; parse new episodes only; streamline Feed.get_episodes
Utility.memoize had an issue where dicts or lists being passed as arguments
would raise an exception. I'm now sanitizing for both before storing the
arguments as a property in the cache object.

Feed.parse will only parse the newest episodes (as the old amount of episodes
are passed to the function, defaulting to 0).

Feed.get_episodes had some weird conditional statement used to determine if
old episodes could be found before. I'm now defaulting the variable to an
empty list if they could not be found -- which streamlines it a bit more and
cleans up the code.